+++ ml/src/pipeline/stages/flow_splitting/__init__.py
@@ -1,1 +1,5 @@
 """Flow-aware workflow entry point splitting stage."""
+
+from pipeline.stages.flow_splitting.main import run
+
+__all__ = ["run"]
