+++ ml/src/pipeline/stages/flow_splitting/constants.py
@@ -1,0 +1,1102 @@
+from __future__ import annotations
+
+CLUSTERS_ARTIFACT = "clusters.json"
+WORKFLOW_ENTRYPOINTS_ARTIFACT = "workflow_entrypoints.json"
+FLOW_SPLIT_REPORT_ARTIFACT = "flow_split_report.json"
+MIN_SPLIT_SIZE = 5
+EXPANDED_MIN_SPLIT_SIZE = 3
+LOW_QUALITY_CLUSTER_DROP_RATIO = 0.80
+NOVEL_REVIEW_CANDIDATE_MIN_SIZE = 3
+FLOW_SPLIT_STRATEGIES = {"conservative", "signal", "expanded"}
+SEQUENCE_SPLIT_PREFIX = "sequence:"
+ACTION_OBJECT_SPLIT_PREFIX = "action_object:"
+ACTION_SPLIT_PREFIX = "action:"
+COMPOUND_SPLIT_SEPARATOR = "|"
+WORKFLOW_HUMAN_REVIEW_CONFIDENCE_THRESHOLD = 0.70
+WORKFLOW_REVIEW_CONFIDENCE_THRESHOLD = WORKFLOW_HUMAN_REVIEW_CONFIDENCE_THRESHOLD
+WORKFLOW_HIGH_CONFIDENCE_THRESHOLD = 0.85
+WORKFLOW_LABEL_BLOCKING_CONFIDENCE_THRESHOLD = 0.50
+WORKFLOW_LABEL_SAMPLE_REVIEW_THRESHOLD = 0.65
+MIN_ENTRYPOINT_DISTINCTIVENESS_FOR_SAMPLE = 0.18
+MIN_ENTRYPOINT_MARGIN_FOR_SAMPLE = 0.0
+MIN_GROUNDED_LABEL_EVIDENCE_COVERAGE = 0.25
+MIN_GROUNDED_TERM_FREQUENCY_SCORE = 0.50
+AUTO_SPLIT_LABEL_EVIDENCE_COVERAGE = 0.45
+AUTO_SPLIT_LABEL_OBJECT_ACTION_JOINT_COVERAGE = 0.65
+AUTO_SPLIT_LABEL_ACTION_OBJECT_VALIDITY = 0.70
+AUTO_SPLIT_LABEL_SPECIFICITY = 0.75
+AUTO_LABEL_MIN_MEMBER_COUNT = 2
+AUTO_REVIEW_ONLY_ACTION_TERMS = frozenset({"정보확인"})
+AUTO_GENERIC_ACTION_TERMS = frozenset({"확인", "가능여부", "가능여부확인"})
+AUTO_MIN_SPECIFIC_OBJECT_LENGTH_FOR_GENERIC_ACTION = 3
+AUTO_WEAK_OBJECT_TERMS = frozenset({"개월", "일", "월", "원", "건", "거", "것", "번", "정도", "부분", "내용"})
+TEXT_INFERRED_ACTION_BLOCKLIST = frozenset({"인증"})
+REVIEW_LABEL_SCAFFOLD_TERMS = frozenset({"문의", "검토", "근거", "부족", "라벨", "추가", "수동"})
+REVIEW_PLACEHOLDER_LABEL = "미분류 문의"
+ACTION_LABEL_HINTS = (
+    "변경",
+    "취소",
+    "해지",
+    "환불",
+    "신청",
+    "조회",
+    "발급",
+    "결제",
+    "청구",
+    "납부",
+    "이체",
+    "출금",
+    "예약",
+    "견적",
+    "연락",
+    "인증",
+    "오류",
+    "문제",
+    "확인",
+    "구매",
+    "가능여부",
+    "정보확인",
+)
+ACTION_INFLECTION_SUFFIXES = (
+    "하다",
+    "하고",
+    "하려고",
+    "할려고",
+    "하려구요",
+    "하려구",
+    "하니까",
+    "니까",
+    "할려구요",
+    "할려구",
+    "할려",
+    "할라",
+    "할래",
+    "할게",
+    "할께",
+    "할",
+    "해",
+    "했",
+    "되",
+    "된",
+    "한",
+)
+FRAME_OBJECT_NOISE_TERMS = frozenset(
+    {
+        "맞습니다",
+        "맞아요",
+        "전화드렸",
+        "전화드렸는데",
+        "그러는데",
+        "들어간",
+        "있을텐데",
+        "상태죠",
+        "체크",
+        "그때",
+        "때는",
+        "대로",
+        "래서",
+        "점이",
+        "나오",
+        "나온",
+        "건지",
+        "어디",
+        "어디로",
+        "어디서",
+        "있다",
+        "하는",
+        "방법",
+        "해주시면",
+        "보내주면",
+        "만약",
+        "없어서",
+        "갈까",
+        "까먹어서",
+        "나가",
+        "나가는",
+        "없이",
+        "저기",
+        "후로",
+        "바꾸",
+        "바꾸었",
+        "바꾸었는지",
+        "바꿔",
+        "것",
+        "것만",
+        "값",
+        "값이",
+        "하려구",
+        "하려구요",
+        "하니까",
+        "확인하니까",
+        "받아서",
+        "받아",
+        "입니다",
+        "문의입니다",
+        "하군",
+        "있을까",
+        "있을까요",
+        "뭐예",
+        "뭐예요",
+        "뭐에요",
+        "뭐지",
+        "뭔지",
+        "요거",
+        "너무",
+        "좋고",
+        "품목별",
+        "간단하게",
+        "한번",
+        "안하게",
+        "주시겠어",
+        "보내주시겠어",
+        "주시고",
+        "나머지",
+        "어저께",
+        "했다는데",
+        "그랬는데",
+        "받았는데",
+        "왔는데",
+        "사용해가지고",
+        "요청해가지고",
+        "그래가지고",
+        "처리했는데",
+        "싶어서",
+        "그걸",
+        "같은데",
+        "새로",
+        "만들",
+        "내고",
+        "예전",
+        "전에",
+        "나간다",
+        "같이",
+        "티비쪽",
+        "있으세",
+        "일은",
+        "쌓이",
+        "들어",
+        "일로",
+        "인가",
+        "사는",
+        "작년",
+        "지난해",
+        "올해",
+        "드는",
+        "될지",
+        "되게끔",
+        "모르겠",
+        "모르겠네",
+        "가능하군",
+        "오르",
+        "오르는",
+        "오르겠",
+        "오르겠죠",
+        "대강",
+        "내야",
+        "아니면",
+        "말",
+        "말을",
+        "말은",
+        "말이",
+        "말이죠",
+        "원래",
+        "오긴",
+        "제대",
+        "어렵다고",
+        "방금",
+        "주십시오",
+        "동의합니다",
+        "고마워",
+        "들어가",
+        "들어가고",
+        "들어간",
+        "카드번호",
+        "전화번호",
+        "휴대폰번호",
+        "핸드폰번호",
+        "계좌번호",
+        "주문번호",
+        "예약번호",
+        "고객번호",
+        "번호",
+        "연락처",
+    }
+)
+ACTION_ONLY_FALLBACK_NOISE_TERMS = FRAME_OBJECT_NOISE_TERMS | frozenset(
+    {
+        "가까운",
+        "정확히",
+        "무조건",
+        "나가",
+        "이번에",
+        "돈이",
+    }
+)
+ACTION_OBJECT_AMBIGUOUS_TERMS = frozenset({"예약", "결제", "연락", "인증"})
+ACTION_TERM_SUPPORT_ALIASES = {
+    "가능여부확인": (
+        "가능",
+        "가능한",
+        "가능한가",
+        "가능한지",
+        "되나요",
+        "될까요",
+        "되는지",
+        "할 수",
+        "할수",
+        "불가",
+        "안 되",
+        "안되",
+        "못 하",
+        "못하",
+    ),
+    "가능여부": (
+        "가능",
+        "가능한",
+        "가능한가",
+        "가능한지",
+        "되나요",
+        "될까요",
+        "되는지",
+        "할 수",
+        "할수",
+        "불가",
+        "안 되",
+        "안되",
+        "못 하",
+        "못하",
+    ),
+    "정보확인": ("정보", "궁금", "알고 싶", "알고싶", "알려", "안내", "문의"),
+    "확인": ("확인", "조회", "상태", "내역", "어디", "언제", "알려"),
+    "변경": ("변경", "바꾸", "수정", "교체"),
+    "취소": ("취소", "철회"),
+    "해지": ("해지", "탈퇴", "중지", "정지", "해제"),
+    "환불": ("환불", "환급", "돌려받", "돌려"),
+    "신청": ("신청", "등록", "접수", "가입"),
+    "발급": ("발급", "재발급"),
+    "결제": ("결제", "납부", "청구", "이체", "입금", "승인"),
+    "청구": ("청구", "결제", "납부"),
+    "납부": ("납부", "결제", "입금"),
+    "이체": ("이체", "입금", "출금"),
+    "출금": ("출금", "인출"),
+    "예약": ("예약", "예매"),
+    "견적": ("견적", "견적서"),
+    "연락": ("연락", "전화", "연결", "상담원", "담당자", "이관"),
+    "인증": ("인증", "본인확인", "본인 확인", "확인"),
+    "오류": ("오류", "장애", "안 되", "안되", "문제"),
+    "문제": ("문제", "오류", "장애", "안 되", "안되"),
+    "구매": ("구매", "구입"),
+}
+LABEL_REVIEW_NOISE_TERMS = ACTION_ONLY_FALLBACK_NOISE_TERMS | frozenset(
+    {
+        "가요",
+        "알게되어서",
+        "드렸는데",
+        "중후한데",
+        "까먹어서",
+        "있잖아",
+        "있잖아요",
+        "품목별",
+        "간단하게",
+        "한번",
+        "안하게",
+        "주시겠어",
+        "보내주시겠어",
+        "주시고",
+        "나머지",
+        "어저께",
+        "했다는데",
+        "그랬는데",
+        "받았는데",
+        "왔는데",
+        "사용해가지고",
+        "요청해가지고",
+        "그래가지고",
+        "처리했는데",
+        "싶어서",
+        "그걸",
+        "같은데",
+        "새로",
+        "만들",
+        "내고",
+        "예전",
+        "전에",
+        "나간다",
+        "같이",
+        "티비쪽",
+        "있으세",
+        "일은",
+        "쌓이",
+        "들어",
+        "일로",
+        "인가",
+        "에게",
+        "보낼",
+        "봐야",
+        "바꾸면",
+        "떨어지",
+        "그래갖고",
+        "해오라",
+        "던데",
+        "싶었던",
+        "곳이라서",
+        "이라서",
+        "라서",
+        "하려면",
+        "궁금해",
+        "좋겠네요",
+        "유효한",
+        "빠른",
+        "바로",
+        "방법",
+        "자세히",
+        "설명",
+        "잠깐",
+        "들어가서",
+        "들어온",
+        "걸로",
+        "해놓",
+        "많았는데",
+        "점이",
+        "생기면",
+        "사게",
+        "쓰다",
+        "쓰다가",
+        "싶었던",
+        "곳이라서",
+        "이렇게",
+    }
+)
+LABEL_RAW_REVIEW_NOISE_SUFFIXES = (
+    "이라서",
+    "라서",
+    "가서",
+    "와서",
+    "어서",
+    "아서",
+    "면서",
+    "는데",
+    "은데",
+    "았는데",
+    "었는데",
+    "어야",
+    "아야",
+    "여야",
+    "려야",
+    "하려면",
+    "다면",
+    "다면요",
+    "이신",
+    "해야",
+    "봐야",
+    "던데",
+    "게",
+    "걸로",
+    "갖고",
+    "떨어지",
+    "싶었던",
+)
+KOREAN_DISCOURSE_LABEL_NOISE_TERMS = frozenset(
+    {
+        "곳인",
+        "마지막",
+        "정말",
+        "부탁드려",
+        "있다면",
+        "것들",
+        "내용들",
+        "구체적",
+        "의논되지",
+        "찍혀",
+        "없죠",
+        "말씀",
+        "말씀이신",
+        "말씀드린",
+        "건데",
+        "안에",
+        "걸린",
+        "거를",
+        "했던",
+        "원을",
+        "그쪽",
+        "내용이나",
+        "상담받고",
+        "가요",
+        "해당",
+        "직접",
+        "뽑아야",
+        "왜그런지",
+        "없더",
+        "옛날",
+        "보내서",
+        "너었으",
+        "그부분",
+        "그라면",
+        "매달",
+        "초에",
+        "넣고",
+        "나서",
+        "나옵",
+        "있어야",
+        "며칠",
+        "기다려야",
+        "할려",
+        "다운시킬",
+        "불러드려야",
+        "걸리나",
+        "빠르다",
+        "문자와서",
+        "뭐죠",
+        "찾았으",
+        "들어가야",
+        "중간",
+        "나왔",
+        "같아",
+        "있다면",
+        "거는",
+        "되는",
+        "되네",
+        "없고",
+        "조금",
+        "비싼",
+        "비싸",
+        "비싸서",
+        "싼",
+        "쓰면",
+        "해요",
+        "거겠네",
+    }
+)
+LABEL_REVIEW_NOISE_TERMS = LABEL_REVIEW_NOISE_TERMS | KOREAN_DISCOURSE_LABEL_NOISE_TERMS
+FRAME_OBJECT_NOISE_TERMS = FRAME_OBJECT_NOISE_TERMS | KOREAN_DISCOURSE_LABEL_NOISE_TERMS
+FLOW_EVENT_LABELS = {
+    "확인질문": "요청확인",
+    "추가정보요청": "정보수집",
+    "정책안내": "기준확인",
+    "불만표현": "문제확인",
+    "예외처리": "예외검토",
+    "해결": "결과안내",
+    "이관": "상담원이관",
+}
+_LABEL_STOPWORDS = frozenset(
+    {
+        "고객",
+        "고객님",
+        "상담",
+        "상담한",
+        "상담하신",
+        "상담하",
+        "상담사",
+        "문의",
+        "문의드",
+        "문의드립니다",
+        "문의드릴",
+        "문의드리",
+        "문의하나드립니다",
+        "문의사항",
+        "확인",
+        "합니다",
+        "있습니다",
+        "그리고",
+        "혹시",
+        "해주세요",
+        "가능",
+        "가능한가요",
+        "가능한지",
+        "가능할까요",
+        "되나요",
+        "되나",
+        "될까요",
+        "처리",
+        "문의해",
+        "그럼",
+        "확인해",
+        "확인해보고",
+        "보겠습니다",
+        "확인해보겠습니다",
+        "저희",
+        "저희가",
+        "그거",
+        "그것",
+        "저거",
+        "거예",
+        "거죠",
+        "이거",
+        "이거죠",
+        "있는",
+        "있나",
+        "없나",
+        "있으면",
+        "되는데",
+        "되죠",
+        "없어",
+        "좋은",
+        "좋네",
+        "좋겠습니다",
+        "남아",
+        "해서",
+        "아이",
+        "오늘",
+        "중에",
+        "여부",
+        "하나",
+        "다른",
+        "바로",
+        "볼게",
+        "연락드릴게",
+        "연락드릴",
+        "드릴게",
+        "드렸어",
+        "해놓게",
+        "전화",
+        "전화로",
+        "통화",
+        "일부터",
+        "일까지",
+        "언제든지",
+        "주세요",
+        "싶은데",
+        "싶어",
+        "정보",
+        "정보들",
+        "품목별",
+        "간단하게",
+        "한번",
+        "안하게",
+        "주시겠어",
+        "보내주시겠어",
+        "주시고",
+        "나머지",
+        "어저께",
+        "했다는데",
+        "그랬는데",
+        "받았는데",
+        "왔는데",
+        "사용해가지고",
+        "요청해가지고",
+        "그래가지고",
+        "처리했는데",
+        "싶어서",
+        "그걸",
+        "같은데",
+        "새로",
+        "만들",
+        "내고",
+        "예전",
+        "전에",
+        "나간다",
+        "같이",
+        "티비쪽",
+        "있으세",
+        "일은",
+        "쌓이",
+        "들어",
+        "일로",
+        "인가",
+        "에게",
+        "보낼",
+        "봐야",
+        "바꾸면",
+        "떨어지",
+        "그래갖고",
+        "해오라",
+        "던데",
+        "싶었던",
+        "곳이라서",
+        "이라서",
+        "라서",
+        "하려면",
+        "곳이",
+        "좋을까",
+        "제가",
+        "저는",
+        "지금",
+        "주세",
+        "그러면",
+        "아마",
+        "그래",
+        "그니깐",
+        "그니까",
+        "그냥",
+        "같은",
+        "알겠습니다",
+        "알았습니다",
+        "알겠어",
+        "그런데",
+        "근데",
+        "하고",
+        "보내고",
+        "보내야",
+        "보내드릴게",
+        "보내드릴게요",
+        "가고",
+        "외에",
+        "그렇군",
+        "그렇군요",
+        "그렇구나",
+        "그렇죠",
+        "그런",
+        "그러",
+        "이제",
+        "어떤",
+        "무슨",
+        "어떻게",
+        "궁금한",
+        "궁금",
+        "궁금해",
+        "좋겠네요",
+        "유효한",
+        "빠른",
+        "바로",
+        "방법",
+        "잠깐",
+        "들어가서",
+        "들어온",
+        "걸로",
+        "해놓",
+        "많았는데",
+        "점이",
+        "생기면",
+        "사게",
+        "쓰다",
+        "쓰다가",
+        "싶었던",
+        "곳이라서",
+        "질문",
+        "생기면",
+        "생기",
+        "여쭤보려고",
+        "여쭤보겠습니다",
+        "여쭤보면",
+        "여쭤",
+        "보려고",
+        "다시",
+        "많이",
+        "되는",
+        "하는데",
+        "있어서",
+        "있으시면",
+        "있습니다",
+        "있어",
+        "아니",
+        "아니라",
+        "이게",
+        "이렇게",
+        "내가",
+        "가지고",
+        "해가지고",
+        "전화해가지고",
+        "필요한",
+        "필요",
+        "정리해서",
+        "으로",
+        "로",
+        "메일",
+        "메일로",
+        "이메일",
+        "대해",
+        "어제",
+        "여기서",
+        "월",
+        "월에",
+        "이번",
+        "진행",
+        "진행하려면",
+        "진행하면",
+        "진행하고",
+        "있어서",
+        "있으면요",
+        "있는지",
+        "적용되지",
+        "포함되지",
+        "않으면",
+        "않는",
+        "생각",
+        "생각하고",
+        "후에",
+        "다양하네",
+        "차이",
+        "어느",
+        "사이",
+        "한지",
+        "주실",
+        "정도",
+        "얼마",
+        "얼마죠",
+        "얼마입니까",
+        "너무",
+        "좋고",
+        "입니까",
+        "뭐예",
+        "뭐예요",
+        "뭐에요",
+        "뭔지",
+        "맞아",
+        "맞아요",
+        "갑자기",
+        "알아봐주세",
+        "알아봐주세요",
+        "감사",
+        "갑사합니다",
+        "고맙습니다",
+        "건강",
+        "고생",
+        "부탁드릴게",
+        "부탁드릴게요",
+        "부탁드립니다",
+        "부탁드려요",
+        "부탁",
+        "뭐야",
+        "뭐니",
+        "보통",
+        "바라겠습니다",
+        "바라겠",
+        "괜찮아",
+        "괜찮은",
+        "괜찮",
+        "괜찮아요",
+        "그게",
+        "아니에",
+        "아니에요",
+        "사항",
+        "수고하세요",
+        "수고했습니다",
+        "수고하셨",
+        "수고하십니다",
+        "수고하십시오",
+        "없으실까",
+        "없으실까요",
+        "없으세",
+        "없습니다",
+        "여보세",
+        "여보세요",
+        "얘기인지",
+        "이어가세요",
+        "이어가십시오",
+        "연결해",
+        "일상",
+        "일상의",
+        "이요",
+        "즐거운",
+        "저기",
+        "여기",
+        "맞습니다",
+        "전화드렸",
+        "전화드렸는데",
+        "그러는데",
+        "들어간",
+        "있을텐데",
+        "상태죠",
+        "체크",
+        "그때",
+        "때는",
+        "대로",
+        "래서",
+        "주말",
+        "잠시만",
+        "드리겠습니다",
+        "보내십시오",
+        "계속",
+        "그래서",
+        "그렇게",
+        "아직",
+        "미정",
+        "상태",
+        "이겠으나",
+        "수와",
+        "일에",
+        "이겠으나",
+        "상태이겠으나",
+        "바꿀",
+        "바꾸",
+        "바꾸는",
+        "하기로",
+        "할게",
+        "할께",
+        "할까",
+        "하겠",
+        "추가로",
+        "해야",
+        "기다릴게",
+        "기다릴게요",
+        "기다리",
+        "받을",
+        "받아",
+        "받는",
+        "받은",
+        "있다고",
+        "왔었는데",
+        "바꾸긴",
+        "잘된",
+        "잘됨",
+        "건가",
+        "거기",
+        "것도",
+        "분",
+        "분은",
+        "분이",
+        "분을",
+        "분한테",
+        "있죠",
+        "드리면",
+        "때문에",
+        "하면",
+        "했어",
+        "됐는데",
+        "거든",
+        "거든요",
+        "쓰고",
+        "있고",
+        "있겠네",
+        "해야되나",
+        "물어",
+        "죄송해",
+        "죄송",
+        "자꾸",
+        "저한테",
+        "가지만",
+        "안낸",
+        "없는데",
+        "제를",
+        "보니",
+        "당연히",
+        "얘기",
+        "나오더라고",
+        "모르고",
+        "있던",
+        "부분이어서",
+        "부분",
+        "것",
+        "것만",
+        "값",
+        "값이",
+        "돼요",
+        "제일",
+        "가서",
+        "들어왔어",
+        "말고",
+        "올려",
+        "하여튼",
+        "항상",
+        "똑같",
+        "어쩌고",
+        "얼마짜리",
+        "따로",
+        "나와",
+        "나온다니까",
+        "곧바",
+        "곧바로",
+        "미리",
+        "있을까",
+        "있을까요",
+        "있는데",
+        "요거",
+        "이런",
+        "저런",
+        "사는",
+        "작년",
+        "지난해",
+        "올해",
+        "엄마",
+        "일단",
+        "에는",
+        "부터",
+        "만약에",
+        "했는데",
+        "그러니까",
+        "하려구",
+        "하려구요",
+        "하니까",
+        "확인하니까",
+        "되고",
+        "받아서",
+        "드는",
+        "될지",
+        "되게끔",
+        "모르겠",
+        "모르겠네",
+        "가능하군",
+        "오르",
+        "오르는",
+        "오르겠",
+        "오르겠죠",
+        "대강",
+        "내야",
+        "아니면",
+        "말",
+        "말을",
+        "말은",
+        "말이",
+        "원래",
+        "오긴",
+        "제대",
+        "어렵다고",
+        "방금",
+        "주십시오",
+        "동의합니다",
+        "고마워",
+        "들어가",
+        "들어가고",
+        "들어간",
+        "했을",
+        "거잖아",
+        "저쩌고",
+        "없이",
+        "까먹어서",
+        "나가",
+        "나가는",
+        "바꾸었",
+        "바꾸었는지",
+        "수가",
+        "얼마나",
+        "해도",
+        "달에",
+        "월요일",
+        "화요일",
+        "수요일",
+        "목요일",
+        "금요일",
+        "토요일",
+        "일요일",
+        "phone_number",
+        "email_address",
+        "date",
+        "time",
+        "name",
+        "number",
+        "birth",
+        "birth_number",
+        "charge",
+        "mobile",
+        "mobile_number",
+        "personal_name",
+        "phone",
+        "email",
+        "address",
+        "account",
+        "amount",
+        "card_number",
+        "order_id",
+        "카드번호",
+        "전화번호",
+        "휴대폰번호",
+        "핸드폰번호",
+        "계좌번호",
+        "주문번호",
+        "예약번호",
+        "고객번호",
+        "번호",
+        "연락처",
+        "요청확인",
+        "기준확인",
+        "정보수집",
+        "결과안내",
+        "상담원이관",
+        "문제확인",
+        "예외검토",
+        "흐름",
+        "미분류",
+        "검토",
+        "문의하나",
+        "들었",
+        "되었어",
+        "연락드리면",
+        "그전",
+        "줄도",
+        "낼라",
+        "안내",
+        "쉬운",
+    }
+)
+_LABEL_STOPWORDS = _LABEL_STOPWORDS | KOREAN_DISCOURSE_LABEL_NOISE_TERMS
+_LABEL_SUFFIXES = (
+    "일부터",
+    "일까지",
+    "부터",
+    "까지",
+    "하려면",
+    "이라서",
+    "라서",
+    "가서",
+    "와서",
+    "어서",
+    "아서",
+    "면서",
+    "는데",
+    "은데",
+    "았는데",
+    "었는데",
+    "어야",
+    "아야",
+    "여야",
+    "려야",
+    "다면",
+    "다면요",
+    "이신",
+    "드린",
+    "건데",
+    "되고",
+    "하면",
+    "싶은데",
+    "싶어서",
+    "싶어",
+    "하려구요",
+    "하려구",
+    "하니까",
+    "니까",
+    "돼서",
+    "받아서",
+    "입니다",
+    "었다고",
+    "었다",
+    "었는데",
+    "다고",
+    "긴",
+    "하십시오",
+    "해보고",
+    "할게",
+    "할까요",
+    "할까",
+    "었는지",
+    "는지",
+    "해야",
+    "게",
+    "걸로",
+    "됐어",
+    "했어",
+    "볼게",
+    "해서",
+    "되죠",
+    "된",
+    "하고",
+    "하",
+    "해",
+    "에는",
+    "에",
+    "으로",
+    "로",
+    "와",
+    "과",
+    "에서",
+    "에게",
+    "의",
+    "은",
+    "는",
+    "이",
+    "가",
+    "을",
+    "를",
+    "도",
+    "만",
+    "할",
+    "거든",
+    "거든요",
+    "됐는데",
+    "인데",
+    "라고",
+    "입니까",
+    "려고",
+    "보려고",
+    "요",
+)
+
